requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.2
billiard==3.6.2.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
django-allauth==0.41.0
django-rq==2.3.0
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rq==1.2.2
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
zipp==3.0.0
